+++ processor/resourcedetectionprocessor/factory.go
@@ -22,6 +22,7 @@ import (
 
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/config"
+	"go.opentelemetry.io/collector/config/confighttp"
 	"go.opentelemetry.io/collector/consumer"
 	"go.opentelemetry.io/collector/processor/processorhelper"
 
@@ -91,13 +92,19 @@ func (*factory) Type() config.Type {
 
 func createDefaultConfig() config.Processor {
 	return &Config{
-		ProcessorSettings: config.NewProcessorSettings(config.NewComponentID(typeStr)),
-		Detectors:         []string{env.TypeStr},
-		Timeout:           5 * time.Second,
-		Override:          true,
+		ProcessorSettings:  config.NewProcessorSettings(config.NewComponentID(typeStr)),
+		Detectors:          []string{env.TypeStr},
+		HTTPClientSettings: defaultHTTPClientSettings(),
+		Override:           true,
 		// TODO: Once issue(https://github.com/open-telemetry/opentelemetry-collector/issues/4001) gets resolved,
 		// 		 Set the default value of 'hostname_source' here instead of 'system' detector
 	}
+}
+
+func defaultHTTPClientSettings() confighttp.HTTPClientSettings {
+	httpClientSettings := confighttp.DefaultHTTPClientSettings()
+	httpClientSettings.Timeout = 5 * time.Second
+	return httpClientSettings
 }
 
 func (f *factory) createTracesProcessor(
@@ -163,14 +170,16 @@ func (f *factory) getResourceDetectionProcessor(
 ) (*resourceDetectionProcessor, error) {
 	oCfg := cfg.(*Config)
 
-	provider, err := f.getResourceProvider(params, cfg.ID(), oCfg.Timeout, oCfg.Detectors, oCfg.DetectorConfig)
+	provider, err := f.getResourceProvider(params, cfg.ID(), oCfg.HTTPClientSettings.Timeout, oCfg.Detectors, oCfg.DetectorConfig)
 	if err != nil {
 		return nil, err
 	}
 
 	return &resourceDetectionProcessor{
-		provider: provider,
-		override: oCfg.Override,
+		provider:           provider,
+		override:           oCfg.Override,
+		httpClientSettings: oCfg.HTTPClientSettings,
+		telemetrySettings:  params.TelemetrySettings,
 	}, nil
 }